services/ai-analyzer/requirements.txt
confluent-kafka
psycopg2-binary
SQLAlchemy
python-dotenv
google-generativeai
